Less ugly
Carousel slides are equal heights and have a dark background to make them stand out from the rest of the page, making things MUCH less jank. The "welcome to Solenoid" header is also on a green background.
TODO: add padding for the next/previous slide buttons, make the slides not be cards because there's no point to it and the rounded corners look bad during transitions. And some more actual functionality.